+++ docs/CHANGELOG.rst
@@ -66,6 +66,8 @@ Changed
 
 Fixed
 -----
+- **SECURITY:** Prevent normal users from creating new ``Processes``
+  over API
 - Configure parallel tests
 - Isolate Elasticsearch indices for parallel tests
 - Fix docker container name for parallel tests